my-app/backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
stripe==14.3.0
pydantic==2.12.5
sendgrid==6.11.0
